+++ assets/asset_names.py
@@ -1,18 +1,18 @@
-import psa_gps.version_vFTrkLn3MMbs9wCLEBBh4s as psa_gps_v1
+import psa_gps.psa_gps_vFTrkLn3MMbs9wCLEBBh4s as psa_gps_v1
 
-import psa_water_sensor_install.version_vuiiHRr2MJSGzFwSncyLP9 as psa_water_sensor_install_v1
-import psa_water_sensor_install.version_vCK5tppVaNkkfWMhtddxEb as psa_water_sensor_install_v2
+import psa_water_sensor_install.wsensor_install_vuiiHRr2MJSGzFwSncyLP9 as psa_water_sensor_install_v1
+import psa_water_sensor_install.wsensor_install_vCK5tppVaNkkfWMhtddxEb as psa_water_sensor_install_v2
 
-import psa_decomp_bag_pre_wt.version_vQnB8sJFc8JEhYJqXiYQRy as psa_decomp_bag_pre_wt_v1
-import psa_decomp_bag_pre_wt.version_vZvVu9PdqRHLeeJGbGha3z as psa_decomp_bag_pre_wt_v2
-import psa_decomp_bag_pre_wt.version_vgaKFRKRg54E2Z7fvayCxf as psa_decomp_bag_pre_wt_v3
-import psa_decomp_bag_pre_wt.version_vzazAbS9satE4PXaGNKwBE as psa_decomp_bag_pre_wt_v4 
+import psa_decomp_bag_pre_wt.decomp_biomass_fresh_vQnB8sJFc8JEhYJqXiYQRy as psa_decomp_bag_pre_wt_v1
+import psa_decomp_bag_pre_wt.decomp_biomass_fresh_vZvVu9PdqRHLeeJGbGha3z as psa_decomp_bag_pre_wt_v2
+import psa_decomp_bag_pre_wt.decomp_biomass_fresh_vgaKFRKRg54E2Z7fvayCxf as psa_decomp_bag_pre_wt_v3
+import psa_decomp_bag_pre_wt.decomp_biomass_fresh_vzazAbS9satE4PXaGNKwBE as psa_decomp_bag_pre_wt_v4 
 
-import psa_bag_dry_wt.version_vDqdEsDav5K6hSRHrgYJmM as psa_bag_dry_wt_v1
-import psa_bag_dry_wt.verision_vusvAjv42DfdfjKkFWm6H3 as psa_bag_dry_wt_v2
+import psa_bag_dry_wt.decomp_biomass_dry_vDqdEsDav5K6hSRHrgYJmM as psa_bag_dry_wt_v1
+import psa_bag_dry_wt.decomp_biomass_dry_vusvAjv42DfdfjKkFWm6H3 as psa_bag_dry_wt_v2
 
-import psa_bag_collect.version_v94kLXWMrZ8fZRpCFzSn4e as psa_bag_collect_v1
-import psa_bag_collect.version_v5LVmcbGVhR3C3fkW9ZhHG as psa_bag_collect_v2
+import psa_bag_collect.decomp_biomass_dry_v94kLXWMrZ8fZRpCFzSn4e as psa_bag_collect_v1
+import psa_bag_collect.decomp_biomass_dry_v5LVmcbGVhR3C3fkW9ZhHG as psa_bag_collect_v2
 
 asset_names = {
     "psa gps": [
